Add files via upload
This project is part of my internship at Oasis Infobyte and represents the completion of Task 1: Customer Segmentation Analysis. The goal of this project is to perform clustering analysis on a dataset to group customers based on shared characteristics. Such insights are valuable for businesses to tailor marketing strategies, improve customer experiences, and boost overall business performance.